requirements.txt
pandas==1.2.3
numpy==1.19.5
scikit-learn==0.24.1
scipy==1.6.1
streamlit==0.78.0
tensorflow==2.4.1
biopython==1.78
bioservices==1.7.11
plotly==4.14.3
xmltodict==0.12.0
